AutoFences Smart Car Android App
• Partnered with Vancouver-based tech-company Mojio and cellular network
provider Telus.
• Adopted Xamarin IDE for both Android and iOS cross-platform porting.
• Established communication with the dongle plugged into the car’s OBD2
diagnostic connector to the Mojio server
• Established interaction between Mojio dongle to Mojio server to Mobile
API.
• Implemented application that allows users to track their car’s data on
the Mojio website or through utilization of applications that manage the
data in novel ways.
• Implemented Speed-Fencing that allows the user to set a speed limit.
If the vehicle travels faster than this speed, the user will receive a
notification.
• Implemented Chrono-Fencing that allows the user to set a time period
in which they would like to be notified if their vehicle starts.
• Implemented Geo-fencing that allows the user to define a region in
which they would like to be notified if their car travels across this
boundary.